# attendance-system/backup_before_restore_20260716_200909/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Attendance System API
  description: >
    API quản lý chấm công tập trung, tích hợp nhiều dòng máy (ZKTeco, Sunbeam/Timmy, Hikvision).
    Hệ thống chỉ lưu trữ raw log chấm công và đồng bộ nhân viên hai chiều — không tính công.
  version: "1.0.0"
servers:
  - url: /api/v1

# JWT Bearer Auth áp dụng cho toàn bộ API (trừ /auth/login được override bên dưới)
security:
  - BearerAuth: []
...
paths:
  /devices:
    post:
      tags: [devices]
      summary: Thêm thiết bị mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceRequest'
      responses:
        '201':
          description: Thiết bị đã được tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
    get:
      tags: [devices]
      summary: Danh sách thiết bị
      responses:
        '200':
          description: Danh sách thiết bị
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
  /devices/{id}:
    put:
      tags: [devices]
      summary: Sửa thiết bị
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceRequest'
      responses:
        '200':
          description: Thiết bị đã được cập nhật
    delete:
      tags: [devices]
      summary: Xoá thiết bị
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      responses:
        '204':
          description: Đã xoá
  /devices/{id}/status:
    get:
      tags: [devices]
      summary: Kiểm tra trạng thái online/offline
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      responses:
        '200':
          description: Trạng thái thiết bị
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceStatus'
  /devices/{id}/test-connection:
    post:
      tags: [devices]
      summary: Test kết nối tới thiết bị
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      responses:
        '200':
          description: Kết nối thành công
        '503':
          description: Không thể kết nối
  /devices/{id}/reboot:
    post:
      tags: [devices]
      summary: Khởi động lại thiết bị (nếu hãng hỗ trợ)
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      responses:
        '202':
          description: Yêu cầu khởi động lại đã được gửi
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Lấy token từ POST /auth/login.
        Gửi kèm header: Authorization: Bearer <token>
  parameters:
    DeviceID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    CreateDeviceRequest:
      type: object
      required: [name, device_type, ip_address, port]
      properties:
        name: { type: string, example: "ZKTeco Kho A" }
        device_type: { type: string, enum: [zkteco, sunbeam, hikvision] }
        ip_address: { type: string, example: "192.168.1.100" }
        port: { type: integer, example: 4370 }
        serial_number: { type: string }
        location: { type: string, example: "Kho hàng tầng 1" }
    UpdateDeviceRequest:
      allOf:
        - $ref: '#/components/schemas/CreateDeviceRequest'
    Device:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        device_type: { type: string }
        ip_address: { type: string }
        port: { type: integer }
        serial_number: { type: string }
        status: { type: string, enum: [online, offline] }
        last_checked_at: { type: string, format: date-time }
        location: { type: string }
        created_at: { type: string, format: date-time }
    DeviceStatus:
      type: object
      properties:
        online: { type: boolean }
        firmware_info: { type: string }
        user_count: { type: integer }
        log_count: { type: integer }
...